Change long to int

+++ debezium-connector-sqlserver/src/main/java/io/debezium/connector/sqlserver/metrics/SqlServerSnapshotPartitionMetrics.java
@@ -56,7 +56,7 @@ class SqlServerSnapshotPartitionMetrics extends AbstractSqlServerPartitionMetric
     }
 
     @Override
-    public long getSnapshotStatusCode() {
+    public int getSnapshotStatusCode() {
         return snapshotMeter.getSnapshotStatusCode();
     }
